Load modals functionality on demand
