+++ src/ggrc/assets/javascripts/entrypoints/dashboard/index.js
@@ -34,9 +34,7 @@ import '../../controllers/tree/tree-view-node';
 import '../../controllers/tree/list_view_controller';
 import '../../controllers/dashboard_widgets_controller';
 import '../../controllers/info_widget_controller';
-import '../../controllers/summary_widget_controller';
 import '../../controllers/dashboard_widget_controller';
-import '../../controllers/modals';
 import '../../controllers/info_pin_controller';
 import '../../controllers/automapper_controller';
 import '../../controllers/dashboard_controller';
